+++ VEngine/Code/Actors/IActorSystem.h
@@ -20,7 +20,7 @@ public:
 	virtual std::vector<Actor*> GetActorsAsBaseClass() = 0;
 	virtual Actor* SpawnActor(const Transform& transform) = 0;
 	virtual Actor* FindActorByName(std::string actorName) = 0;
-	virtual uint32_t GetNumActors() = 0;
+	virtual size_t GetNumActors() = 0;
 	virtual void DeferActorForDestroy(int index) = 0;
 	virtual void DestroyDeferredActors() = 0;
